pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "paycalc"
version = "0.2.0"
description = "Personal pay and tax projection tools - SDK, CLI, and MCP server"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "PyPDF2>=3.0.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
mcp = [
    "mcp>=1.0.0",
    "pydantic>=2.0.0",
]
filter = [
    "jsonpath-ng>=1.6.0",
]
dev = [
    "pytest>=7.0",
    "ruff>=0.1.0",
]

[project.scripts]
pay-calc = "paycalc.cli.__main__:main"
pay-calc-mcp = "paycalc.mcp.server:run_server"

[tool.setuptools.packages.find]
where = ["."]
include = ["paycalc*", "processors*"]

[tool.setuptools.package-data]
paycalc = ["schemas/*.json", "config/*.yaml"]
"*" = ["tax-rules/*.yaml"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "W",   # pycodestyle warnings
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
]
ignore = [
    "E501", # Line too long (handled by formatter)
]
